+++ federation/keyserver.go
@@ -8,12 +8,12 @@ package federation
 
 import (
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"strconv"
 	"time"
 
 	"github.com/gorilla/mux"
+	"go.mau.fi/util/exhttp"
 	"go.mau.fi/util/jsontime"
 
 	"maunium.net/go/mautrix"
@@ -47,6 +47,7 @@ type KeyServer struct {
 	KeyProvider     ServerKeyProvider
 	Version         ServerVersion
 	WellKnownTarget string
+	OtherKeys       KeyCache
 }
 
 // Register registers the key server endpoints to the given router.
@@ -58,23 +59,11 @@ func (ks *KeyServer) Register(r *mux.Router) {
 	keyRouter.HandleFunc("/v2/query/{serverName}", ks.GetQueryKeys).Methods(http.MethodGet)
 	keyRouter.HandleFunc("/v2/query", ks.PostQueryKeys).Methods(http.MethodPost)
 	keyRouter.NotFoundHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		jsonResponse(w, http.StatusNotFound, &mautrix.RespError{
-			ErrCode: mautrix.MUnrecognized.ErrCode,
-			Err:     "Unrecognized endpoint",
-		})
+		mautrix.MUnrecognized.WithStatus(http.StatusNotFound).WithMessage("Unrecognized endpoint").Write(w)
 	})
 	keyRouter.MethodNotAllowedHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		jsonResponse(w, http.StatusMethodNotAllowed, &mautrix.RespError{
-			ErrCode: mautrix.MUnrecognized.ErrCode,
-			Err:     "Invalid method for endpoint",
-		})
+		mautrix.MUnrecognized.WithStatus(http.StatusMethodNotAllowed).WithMessage("Invalid method for endpoint").Write(w)
 	})
-}
-
-func jsonResponse(w http.ResponseWriter, code int, data any) {
-	w.Header().Add("Content-Type", "application/json")
-	w.WriteHeader(code)
-	_ = json.NewEncoder(w).Encode(data)
 }
 
 // RespWellKnown is the response body for the `GET /.well-known/matrix/server` endpoint.
@@ -87,12 +76,9 @@ type RespWellKnown struct {
 // https://spec.matrix.org/v1.9/server-server-api/#get_well-knownmatrixserver
 func (ks *KeyServer) GetWellKnown(w http.ResponseWriter, r *http.Request) {
 	if ks.WellKnownTarget == "" {
-		jsonResponse(w, http.StatusNotFound, &mautrix.RespError{
-			ErrCode: mautrix.MNotFound.ErrCode,
-			Err:     "No well-known target set",
-		})
+		mautrix.MNotFound.WithMessage("No well-known target set").Write(w)
 	} else {
-		jsonResponse(w, http.StatusOK, &RespWellKnown{Server: ks.WellKnownTarget})
+		exhttp.WriteJSONResponse(w, http.StatusOK, &RespWellKnown{Server: ks.WellKnownTarget})
 	}
 }
 
@@ -105,7 +91,7 @@ type RespServerVersion struct {
 //
 // https://spec.matrix.org/v1.9/server-server-api/#get_matrixfederationv1version
 func (ks *KeyServer) GetServerVersion(w http.ResponseWriter, r *http.Request) {
-	jsonResponse(w, http.StatusOK, &RespServerVersion{Server: ks.Version})
+	exhttp.WriteJSONResponse(w, http.StatusOK, &RespServerVersion{Server: ks.Version})
 }
 
 // GetServerKey implements the `GET /_matrix/key/v2/server` endpoint.
@@ -114,12 +100,9 @@ func (ks *KeyServer) GetServerVersion(w http.ResponseWriter, r *http.Request) {
 func (ks *KeyServer) GetServerKey(w http.ResponseWriter, r *http.Request) {
 	domain, key := ks.KeyProvider.Get(r)
 	if key == nil {
-		jsonResponse(w, http.StatusNotFound, &mautrix.RespError{
-			ErrCode: mautrix.MNotFound.ErrCode,
-			Err:     fmt.Sprintf("No signing key found for %q", r.Host),
-		})
+		mautrix.MNotFound.WithMessage("No signing key found for %q", r.Host).Write(w)
 	} else {
-		jsonResponse(w, http.StatusOK, key.GenerateKeyResponse(domain, nil))
+		exhttp.WriteJSONResponse(w, http.StatusOK, key.GenerateKeyResponse(domain, nil))
 	}
 }
 
@@ -144,10 +127,7 @@ func (ks *KeyServer) PostQueryKeys(w http.ResponseWriter, r *http.Request) {
 	var req ReqQueryKeys
 	err := json.NewDecoder(r.Body).Decode(&req)
 	if err != nil {
-		jsonResponse(w, http.StatusBadRequest, &mautrix.RespError{
-			ErrCode: mautrix.MBadJSON.ErrCode,
-			Err:     fmt.Sprintf("failed to parse request: %v", err),
-		})
+		mautrix.MBadJSON.WithMessage("failed to parse request: %v", err).Write(w)
 		return
 	}
 
@@ -165,7 +145,7 @@ func (ks *KeyServer) PostQueryKeys(w http.ResponseWriter, r *http.Request) {
 			}
 		}
 	}
-	jsonResponse(w, http.StatusOK, resp)
+	exhttp.WriteJSONResponse(w, http.StatusOK, resp)
 }
 
 // GetQueryKeysResponse is the response body for the `GET /_matrix/key/v2/query/{serverName}` endpoint
@@ -181,23 +161,35 @@ func (ks *KeyServer) GetQueryKeys(w http.ResponseWriter, r *http.Request) {
 	minimumValidUntilTSString := r.URL.Query().Get("minimum_valid_until_ts")
 	minimumValidUntilTS, err := strconv.ParseInt(minimumValidUntilTSString, 10, 64)
 	if err != nil && minimumValidUntilTSString != "" {
-		jsonResponse(w, http.StatusBadRequest, &mautrix.RespError{
-			ErrCode: mautrix.MInvalidParam.ErrCode,
-			Err:     fmt.Sprintf("failed to parse ?minimum_valid_until_ts: %v", err),
-		})
+		mautrix.MInvalidParam.WithMessage("failed to parse ?minimum_valid_until_ts: %v", err).Write(w)
 		return
 	} else if time.UnixMilli(minimumValidUntilTS).After(time.Now().Add(24 * time.Hour)) {
-		jsonResponse(w, http.StatusBadRequest, &mautrix.RespError{
-			ErrCode: mautrix.MInvalidParam.ErrCode,
-			Err:     "minimum_valid_until_ts may not be more than 24 hours in the future",
-		})
+		mautrix.MInvalidParam.WithMessage("minimum_valid_until_ts may not be more than 24 hours in the future").Write(w)
 		return
 	}
 	resp := &GetQueryKeysResponse{
 		ServerKeys: []*ServerKeyResponse{},
 	}
-	if domain, key := ks.KeyProvider.Get(r); key != nil && domain == serverName {
-		resp.ServerKeys = append(resp.ServerKeys, key.GenerateKeyResponse(serverName, nil))
+	domain, key := ks.KeyProvider.Get(r)
+	if domain == serverName {
+		if key != nil {
+			resp.ServerKeys = append(resp.ServerKeys, key.GenerateKeyResponse(serverName, nil))
+		}
+	} else if ks.OtherKeys != nil {
+		otherKey, err := ks.OtherKeys.LoadKeys(serverName)
+		if err != nil {
+			mautrix.MUnknown.WithMessage("Failed to load keys from cache").Write(w)
+			return
+		}
+		if key != nil && domain != "" {
+			signature, err := key.SignJSON(otherKey)
+			if err == nil {
+				otherKey.Signatures[domain] = map[id.KeyID]string{
+					key.ID: signature,
+				}
+			}
+		}
+		resp.ServerKeys = append(resp.ServerKeys, otherKey)
 	}
-	jsonResponse(w, http.StatusOK, resp)
+	exhttp.WriteJSONResponse(w, http.StatusOK, resp)
 }